+++ scripts/00.fonts.sh
@@ -1,4 +1,8 @@
 # fonts
+if ! exists brew; then
+  return
+fi
+
 install-fonts() {
   info "Installing fonts..."
   brew install --cask font-monaspace
