+++ scripts/00.zed.sh
@@ -1,4 +1,8 @@
 # ZED (A modern text editor): https://zed.dev/
+if ! exists brew; then
+  return
+fi
+
 install-zed() {
   info "Installing zed..."
   brew install --cask zed
