+++ scripts/00.zig.sh
@@ -1,4 +1,8 @@
 # ZIG (programming language): https://ziglang.org/
+if ! exists brew; then
+  return
+fi
+
 install-zig() {
   info "Installing zig..."
   brew install zig
